+++ HISTORY.rst
@@ -3,6 +3,14 @@
 Release History
 ---------------
 
+2.1.0 (2014-xx-xx)
+++++++++++++++++++
+- Allow keyword arguments in functions: `octave.plot([1,2,3], linewidth=2))`
+  These are translated to ("prop", value) arguments to the function.
+- Add option to show plotting gui with `-g` flag in OctaveMagic.
+- Add ability to specify the Octave executable as a keyword argument to
+  the Oct2Py object.
+
 2.0.0 (2014-08-14)
 ++++++++++++++++++
 - **Breaking changes**